fix(WiCan): uses GPIO_NUM_NC(-1) for non-assigned LED; renames LEDs to new scheme

+++ main/comm_server.h
@@ -21,7 +21,7 @@
 
 #ifndef __COMM_SERVER_H__
 #define __COMM_SERVER_H__
-int8_t tcp_server_init(uint32_t port, QueueHandle_t *xTXp_Queue, QueueHandle_t *xRXp_Queue, uint8_t connected_led, uint8_t udp_en);
+int8_t tcp_server_init(uint32_t port, QueueHandle_t *xTXp_Queue, QueueHandle_t *xRXp_Queue, int connected_led, uint8_t udp_en);
 int8_t tcp_port_open(void);
 
 void tcp_server_suspend(void);
